working on offline part

+++ app/src/main/java/com/example/meeera/browserapp/Fragment/News.kt
@@ -1,8 +1,10 @@
 package com.example.meeera.browserapp.Fragment
 
 import android.content.Context
+import android.net.ConnectivityManager
 import android.os.Bundle
 import android.support.v4.app.Fragment
+import android.support.v4.widget.SwipeRefreshLayout
 import android.support.v7.widget.LinearLayoutManager
 import android.support.v7.widget.RecyclerView
 import android.util.Log
@@ -11,7 +13,9 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.meeera.browserapp.Adapter.NewsAdapter
 import com.example.meeera.browserapp.Model.ArticleDetail
+import com.example.meeera.browserapp.MyApplication
 import com.example.meeera.browserapp.R
+import com.example.meeera.browserapp.Service.ConnectivityReceiver
 import com.example.meeera.browserapp.Service.NetworkCall
 import io.realm.Realm
 import java.text.SimpleDateFormat
@@ -26,6 +30,7 @@ class News : Fragment(), NetworkCall.getData {
     var realm: Realm by Delegates.notNull()
     var recyclerview: RecyclerView? = null
     var flag: Boolean = false
+    var swipeToRefresh : SwipeRefreshLayout ?= null
     override fun onAttach(context: Context?) {
         super.onAttach(context)
         realm = Realm.getDefaultInstance()
@@ -37,7 +42,7 @@ class News : Fragment(), NetworkCall.getData {
             if (details[0].publishedAt?.substring(0, 10) == df.format(calendar.time)) {
                 flag = true
             }
-            if (!flag) {
+            if (!flag && isConnected()) {
                 realm.executeTransaction {
                     realm.where(ArticleDetail::class.java).findAll().deleteAllFromRealm()
                 }
@@ -51,11 +56,22 @@ class News : Fragment(), NetworkCall.getData {
     override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         var view = inflater?.inflate(R.layout.news_fragment, container, false)
         recyclerview = view?.findViewById(R.id.grdlist) as RecyclerView
-        if (flag) {
+        swipeToRefresh = view?.findViewById(R.id.swipeToRefresh) as SwipeRefreshLayout
+        Log.d("connected", "connect"+ isConnected())
+        if (flag || !isConnected()) {
             Log.d("amit", "flag" + flag)
             recyclerview?.adapter = NewsAdapter(activity, realm.where(ArticleDetail::class.java).findAll(), true)
             recyclerview?.layoutManager = LinearLayoutManager(context)
         }
+        swipeToRefresh?.setOnRefreshListener({
+            if(isConnected()) {
+                realm.executeTransaction {
+                    realm.where(ArticleDetail::class.java).findAll().deleteAllFromRealm()
+                }
+                NetworkCall(this).getSub()
+            }
+            swipeToRefresh?.isRefreshing = false
+        })
         return view
     }
 
@@ -75,4 +91,7 @@ class News : Fragment(), NetworkCall.getData {
         recyclerview?.layoutManager = LinearLayoutManager(context)
     }
 
+    fun isConnected() : Boolean {
+        return ConnectivityReceiver.isConnected()
+    }
 }
